add latency_dataset config files, move wrappermodel classes, update latency_dataset.py
